Option to turn off access to remote Data Archive during start-up
