+++ pom.xml
@@ -26,7 +26,7 @@
     </parent>
 
     <artifactId>dd-data-vault</artifactId>
-    <version>4.1.1-SNAPSHOT</version>
+    <version>4.1.3-SNAPSHOT</version>
 
     <name>DD Data Vault</name>
     <url>https://github.com/DANS-KNAW/dd-data-vault</url>
@@ -38,7 +38,7 @@
         <commons-validator.version>1.7</commons-validator.version>
         <dans-ocfl-extensions.version>1.1.0</dans-ocfl-extensions.version>
         <dans-ocfl-java-extensions-lib.version>2.0.0</dans-ocfl-java-extensions-lib.version>
-        <dans-layer-store-lib.version>2.0.0</dans-layer-store-lib.version>
+        <dans-layer-store-lib.version>2.1.0</dans-layer-store-lib.version>
         <dd-data-vault-api.version>1.1.1-SNAPSHOT</dd-data-vault-api.version>
         <dans-java-utils.version>3.0.0</dans-java-utils.version>
         <main-class>nl.knaw.dans.datavault.DdDataVaultApplication</main-class>
@@ -85,7 +85,6 @@
         <dependency>
             <groupId>io.ocfl</groupId>
             <artifactId>ocfl-java-core</artifactId>
-            <version>2.2.3</version>
         </dependency>
         <dependency>
             <groupId>nl.knaw.dans</groupId>
